bossphorus/pyproject.toml
[tool.poetry]
name = "bossphorus"
version = "1.0.0"
description = "Baby bossDB"
authors = ["Jordan Matelsky <jordan.matelsky@jhuapl.edu>", "Miller Wilt <miller.wilt@jhuapl.edu>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.0"
blosc = "^1.6"
typing = "^3.6"
typing-extensions = "^3.6"
pyzmq = "^17.1"
attrs = "^18.2"
numpy = "^1.15"

[tool.poetry.dev-dependencies]
pylint = "^2.1"
pytest = "^4.0"
mypy = "^0.641.0"
autopep8 = "^1.4"
rope = "^0.11.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
